Move tests to samples as integration so junit is happy
