+++ moshi-sealed/sealed-interfaces-samples/java/src/main/java/dev/zacsweers/moshix/sealed/sample/java/MessageWithNullDefault.java
@@ -4,18 +4,21 @@ import com.squareup.moshi.JsonClass;
 
 import java.util.Map;
 
+import dev.zacsweers.moshix.sealed.annotations.DefaultNull;
 import dev.zacsweers.moshix.sealed.annotations.TypeLabel;
 
 // @DefaultObject is not possible in java
+@DefaultNull
 @JsonClass(generateAdapter = true, generator = "sealed:type")
-sealed interface MessageInterface
-    permits MessageInterface.Success, MessageInterface.Error {
+sealed interface MessageWithNullDefault
+    permits MessageWithNullDefault.Success, MessageWithNullDefault.Error {
 
   @TypeLabel(label = "success", alternateLabels = {"successful"})
-  final record Success(String value) implements MessageInterface {
+  final record Success(String value) implements MessageWithNullDefault {
   }
 
   @TypeLabel(label = "error")
-  final record Error(Map<String, Object> error_logs) implements MessageInterface {
+  final record Error(Map<String, Object> error_logs) implements
+      MessageWithNullDefault {
   }
 }
